[iOS] Images are displayed for all speakers (where available)
